+++ lib/app/connectivity_model.dart
@@ -1,10 +1,17 @@
 import 'dart:async';
 
 import 'package:connectivity_plus/connectivity_plus.dart';
+import 'package:flutter/material.dart';
 import 'package:safe_change_notifier/safe_change_notifier.dart';
+import 'package:watch_it/watch_it.dart';
 
 import '../common/data/audio_type.dart';
+import '../common/view/snackbars.dart';
+import '../extensions/connectivity_x.dart';
+import '../l10n/l10n.dart';
+import '../player/player_model.dart';
 import '../player/player_service.dart';
+import '../settings/settings_model.dart';
 
 class ConnectivityModel extends SafeChangeNotifier {
   ConnectivityModel({
@@ -15,6 +22,8 @@ class ConnectivityModel extends SafeChangeNotifier {
 
   final PlayerService _playerService;
   final Connectivity _connectivity;
+  Stream<List<ConnectivityResult>> get onConnectivityChanged =>
+      _connectivity.onConnectivityChanged;
   StreamSubscription<List<ConnectivityResult>>? _connectivitySubscription;
 
   Future<void> init() async {
@@ -34,6 +43,9 @@ class ConnectivityModel extends SafeChangeNotifier {
 
   bool get isOnline => _connectivity.isOnline(_result);
 
+  bool get isMaybeLowBandWidth => _connectivity.isNotWifiNorEthernet(_result);
+
+  List<ConnectivityResult>? get result => _result;
   List<ConnectivityResult>? _result;
   void _updateConnectivity(List<ConnectivityResult> newResult) {
     if (!_connectivity.isOnline(newResult) &&
@@ -51,11 +63,30 @@ class ConnectivityModel extends SafeChangeNotifier {
   }
 }
 
-extension _ConnectivityX on Connectivity {
-  bool isOnline(List<ConnectivityResult>? res) =>
-      res?.contains(ConnectivityResult.ethernet) == true ||
-      res?.contains(ConnectivityResult.bluetooth) == true ||
-      res?.contains(ConnectivityResult.mobile) == true ||
-      res?.contains(ConnectivityResult.vpn) == true ||
-      res?.contains(ConnectivityResult.wifi) == true;
+void onConnectivityChangedHandler(
+  BuildContext context,
+  AsyncSnapshot<List<ConnectivityResult>?> res,
+  void Function() cancel,
+) {
+  final l10n = context.l10n;
+  final dataSafeMode = di<PlayerModel>().dataSafeMode;
+  final notifyDataSafeMode = di<SettingsModel>().notifyDataSafeMode;
+  if (!res.hasData || !context.mounted || !notifyDataSafeMode) {
+    return;
+  }
+
+  if (!dataSafeMode && di<Connectivity>().isNotWifiNorEthernet(res.data)) {
+    di<PlayerModel>().setDataSafeMode(true);
+    showSnackBar(
+      context: context,
+      content: Text(l10n.dataSafeModeEnabled),
+    );
+  } else if (dataSafeMode &&
+      !di<Connectivity>().isNotWifiNorEthernet(res.data)) {
+    di<PlayerModel>().setDataSafeMode(false);
+    showSnackBar(
+      context: context,
+      content: Text(l10n.dataSafeModeDisabled),
+    );
+  }
 }