+++ lib/app/view/mobile_page.dart
@@ -1,7 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:watch_it/watch_it.dart';
 
-import '../../common/view/snackbars.dart';
 import '../../common/view/theme.dart';
 import '../../extensions/build_context_x.dart';
 import '../../player/player_model.dart';
@@ -9,9 +8,8 @@ import '../../player/view/full_height_player.dart';
 import '../../player/view/player_view.dart';
 import '../../podcasts/download_model.dart';
 import '../../podcasts/podcast_model.dart';
-import '../../podcasts/podcast_search_state.dart';
-import '../../podcasts/view/podcast_snackbar_contents.dart';
 import '../app_model.dart';
+import '../connectivity_model.dart';
 import 'mobile_bottom_bar.dart';
 
 class MobilePage extends StatelessWidget with WatchItMixin {
@@ -29,41 +27,17 @@ class MobilePage extends StatelessWidget with WatchItMixin {
 
     registerStreamHandler(
       select: (DownloadModel m) => m.messageStream,
-      initialValue: null,
-      handler: (context, snapshot, cancel) {
-        if (snapshot.hasData) {
-          showSnackBar(context: context, content: Text(snapshot.data ?? ''));
-        }
-      },
+      handler: downloadMessageStreamHandler,
     );
 
     registerStreamHandler(
       select: (PodcastModel m) => m.stateStream,
-      initialValue: null,
-      handler: (context, newValue, cancel) {
-        if (newValue.hasData) {
-          if (newValue.data == PodcastSearchState.done) {
-            ScaffoldMessenger.of(context).clearSnackBars();
-          } else {
-            showSnackBar(
-              context: context,
-              content: switch (newValue.data) {
-                PodcastSearchState.loading =>
-                  const PodcastSearchLoadingSnackBarContent(),
-                PodcastSearchState.empty =>
-                  const PodcastSearchEmptyFeedSnackBarContent(),
-                PodcastSearchState.timeout =>
-                  const PodcastSearchTimeoutSnackBarContent(),
-                _ => const SizedBox.shrink()
-              },
-              duration: switch (newValue.data) {
-                PodcastSearchState.loading => const Duration(seconds: 1000),
-                _ => const Duration(seconds: 3),
-              },
-            );
-          }
-        }
-      },
+      handler: podcastStateStreamHandler,
+    );
+
+    registerStreamHandler(
+      select: (ConnectivityModel m) => m.onConnectivityChanged,
+      handler: onConnectivityChangedHandler,
     );
 
     return Scaffold(