+++ lib/player/view/player_main_controls.dart
@@ -9,9 +9,9 @@ import '../../common/view/theme.dart';
 import '../../extensions/build_context_x.dart';
 import '../../extensions/theme_data_x.dart';
 import '../../l10n/l10n.dart';
+import '../../radio/view/next_station_button.dart';
 import '../player_model.dart';
 import 'play_button.dart';
-import 'player_pause_timer_button.dart';
 import 'repeat_button.dart';
 import 'seek_button.dart';
 import 'shuffle_button.dart';
@@ -117,7 +117,7 @@ class PlayerMainControls extends StatelessWidget with WatchItMixin {
             active: active,
             iconColor: defaultColor,
           ),
-        AudioType.radio => const PlayerPauseTimerButton(),
+        AudioType.radio => const NextStationButton(),
         _ => const SizedBox.shrink(),
       },
     ];