+++ lib/podcasts/view/podcasts_collection_body.dart
@@ -5,6 +5,7 @@ import 'package:watch_it/watch_it.dart';
 import '../../app/connectivity_model.dart';
 import '../../common/data/audio.dart';
 import '../../common/data/audio_type.dart';
+import '../../common/page_ids.dart';
 import '../../common/view/adaptive_container.dart';
 import '../../common/view/audio_card.dart';
 import '../../common/view/audio_card_bottom.dart';
@@ -14,7 +15,6 @@ import '../../common/view/no_search_result_page.dart';
 import '../../common/view/offline_page.dart';
 import '../../common/view/safe_network_image.dart';
 import '../../common/view/theme.dart';
-import '../../constants.dart';
 import '../../extensions/build_context_x.dart';
 import '../../l10n/l10n.dart';
 import '../../library/library_model.dart';
@@ -37,7 +37,6 @@ class PodcastsCollectionBody extends StatelessWidget with WatchItMixin {
     final loading =
         watchPropertyValue((PodcastModel m) => m.checkingForUpdates);
     final subs = watchPropertyValue((LibraryModel m) => m.podcasts);
-    watchPropertyValue((LibraryModel m) => m.podcastUpdatesLength);
     final libraryModel = di<LibraryModel>();
     final updatesLength =
         watchPropertyValue((LibraryModel m) => m.podcastUpdatesLength);
@@ -64,7 +63,7 @@ class PodcastsCollectionBody extends StatelessWidget with WatchItMixin {
             ),
             ImportantButton(
               onPressed: () {
-                di<LibraryModel>().push(pageId: kSearchPageId);
+                di<LibraryModel>().push(pageId: PageIDs.searchPage);
                 di<SearchModel>()
                   ..setAudioType(AudioType.podcast)
                   ..setSearchQuery(null)
@@ -107,19 +106,23 @@ class PodcastsCollectionBody extends StatelessWidget with WatchItMixin {
                     itemCount: itemCount,
                     gridDelegate: audioCardGridDelegate,
                     itemBuilder: (context, index) {
-                      final MapEntry<String, List<Audio>> podcast;
+                      final MapEntry<String, List<Audio>>? podcast;
                       if (updatesOnly) {
                         podcast = subs.entries
                             .where(
                               (e) => libraryModel.podcastUpdateAvailable(e.key),
                             )
-                            .elementAt(index);
+                            .elementAtOrNull(index);
                       } else if (downloadsOnly) {
                         podcast = subs.entries
                             .where((e) => libraryModel.feedHasDownload(e.key))
-                            .elementAt(index);
+                            .elementAtOrNull(index);
                       } else {
                         podcast = subs.entries.elementAt(index);
+                      }
+
+                      if (podcast == null) {
+                        return const SizedBox.shrink();
                       }
 
                       final artworkUrl600 =
@@ -152,15 +155,15 @@ class PodcastsCollectionBody extends StatelessWidget with WatchItMixin {
                         ),
                         onPlay: () => di<PlayerModel>()
                             .startPlaylist(
-                              audios: podcast.value,
+                              audios: podcast!.value,
                               listName: podcast.key,
                             )
                             .then(
                               (_) => libraryModel.removePodcastUpdate(
-                                podcast.key,
+                                podcast!.key,
                               ),
                             ),
-                        onTap: () => libraryModel.push(pageId: podcast.key),
+                        onTap: () => libraryModel.push(pageId: podcast!.key),
                       );
                     },
                   ),